chore: updated seeding to a trpc router to trigger from frontend
