+++ payload-types.ts
@@ -32,7 +32,7 @@ export interface User {
   id: string;
   name?: string | null;
   imageUrl?: string | null;
-  role?: ('admin' | 'user') | null;
+  role?: ('admin' | 'user' | 'author') | null;
   emailVerified?: string | null;
   accounts?:
     | {
