+++ payload.config.ts
@@ -32,9 +32,6 @@ import {
   generateURL,
 } from '@/utils/seo'
 
-import { seed } from './src/payload/seed'
-import { blogPost } from './src/payload/seed/data/blog'
-
 // import {
 //   generateDescription,
 //   generateImage,
@@ -98,169 +95,6 @@ export default buildConfig({
   csrf: [env.NEXT_PUBLIC_PUBLIC_URL],
   collections: [Users, Media, Blogs, Pages, Sessions],
   globals: [siteSettings],
-  async onInit(payload) {
-    const blogImagePath = [
-      {
-        data: { alt: 'blog image-1' },
-        options: {
-          filePath: './media/seed/blog-1.jpg',
-        },
-      },
-      {
-        data: { alt: 'blog image-2' },
-        options: {
-          filePath: './media/seed/blog-2.jpg',
-        },
-      },
-      {
-        data: { alt: 'blog image-3' },
-        options: {
-          filePath: './media/seed/blog-3.jpg',
-        },
-      },
-      {
-        data: { alt: 'blog image-4' },
-        options: {
-          filePath: './media/seed/blog-4.jpg',
-        },
-      },
-      {
-        data: { alt: 'blog image-5' },
-        options: {
-          filePath: './media/seed/blog-5.jpg',
-        },
-      },
-      {
-        data: { alt: 'blog image-6' },
-        options: {
-          filePath: './media/seed/blog-6.jpg',
-        },
-      },
-      {
-        data: { alt: 'blog image-6' },
-        options: {
-          filePath: './media/seed/blog-7.jpg',
-        },
-      },
-      {
-        data: { alt: 'blog image-7' },
-        options: {
-          filePath: './media/seed/blog-8.jpg',
-        },
-      },
-    ]
-
-    const authorImagePath = [
-      {
-        data: { alt: 'author image-1' },
-        options: {
-          filePath: './media/seed/blogAuthor-1.jpg',
-        },
-      },
-      {
-        data: { alt: 'author image-2' },
-        options: {
-          filePath: './media/seed/blogAuthor-2.jpg',
-        },
-      },
-      {
-        data: { alt: 'author image-3' },
-        options: {
-          filePath: './media/seed/blogAuthor-3.jpg',
-        },
-      },
-      {
-        data: { alt: 'author image-4' },
-        options: {
-          filePath: './media/seed/blogAuthor-4.jpg',
-        },
-      },
-      {
-        data: { alt: 'author image-5' },
-        options: {
-          filePath: './media/seed/blogAuthor-5.jpg',
-        },
-      },
-      {
-        data: { alt: 'author image-6' },
-        options: {
-          filePath: './media/seed/blogAuthor-6.jpg',
-        },
-      },
-      {
-        data: { alt: 'author image-6' },
-        options: {
-          filePath: './media/seed/blogAuthor-7.jpg',
-        },
-      },
-      {
-        data: { alt: 'author image-7' },
-        options: {
-          filePath: './media/seed/blogAuthor-8.jpg',
-        },
-      },
-    ]
-
-    const blogImages: any = await seed({
-      payload,
-      collectionsToSeed: [
-        {
-          collectionSlug: 'media',
-          seed: [...blogImagePath],
-        },
-      ],
-    })
-
-    if (
-      blogImages?.at(0)?.status === 'fulfilled' &&
-      !blogImages?.at(0)?.value?.result?.message
-    ) {
-      const authorImages: any = await seed({
-        payload,
-        collectionsToSeed: [
-          {
-            collectionSlug: 'media',
-            seed: [...authorImagePath],
-          },
-        ],
-        skipSeeding: false,
-      })
-    }
-
-    if (
-      blogImages?.at(0)?.status === 'fulfilled' &&
-      !blogImages?.at(0)?.value?.result?.message &&
-      Array.isArray(blogImages?.at(0)?.value?.result) &&
-      authorImages?.at(0)?.status === 'fulfilled' &&
-      !authorImages?.at(0)?.value?.result?.message &&
-      Array.isArray(authorImages?.at(0)?.value?.result)
-    ) {
-      const seedData = blogPost.map((blogPost, index) => {
-        const blogImage = blogImages?.at(0)?.value?.result.at(index)
-        const authorImage = authorImages?.at(0)?.value?.result.at(index)
-
-        return {
-          data: {
-            ...blogPost,
-            blog_image:
-              blogImage.status === 'fulfilled' ? blogImage?.value?.id : '',
-            authorImage:
-              authorImage.status === 'fulfilled' ? authorImage?.value?.id : '',
-          },
-        }
-      })
-
-      const result = await seed({
-        payload,
-        collectionsToSeed: [
-          {
-            collectionSlug: 'blogs',
-            seed: [...seedData],
-          },
-        ],
-      })
-    }
-  },
   plugins: [
     nestedDocsPlugin({
       collections: [COLLECTION_SLUG_PAGE],
