+++ src/trpc/router/index.ts
@@ -1,6 +1,7 @@
 import { publicProcedure, router } from '@/trpc'
 import { getBlogs } from '@/trpc/router/blog-router'
 import { getLayouts } from '@/trpc/router/page-router'
+import { seed } from '@/trpc/router/seed'
 import { getSiteSettings } from '@/trpc/router/site-settings-router'
 import { todoRouter } from '@/trpc/router/todo'
 
@@ -9,6 +10,7 @@ export const appRouter = router({
   page: getLayouts,
   blog: getBlogs,
   SiteSettings: getSiteSettings,
+  seed: seed,
   test: publicProcedure.query(async () => {
     return {
       success: 'working',
